+++ src/br/com/solstice/notecommerce/dao/impl/domain/product/ProductDAO.java
@@ -25,7 +25,7 @@ public class ProductDAO extends AbstractDAO {
 	}
 
 	@Override
-	public int save(Entity entity) {
+	public long save(Entity entity) {
 		openConnection();
 		PreparedStatement pstm = null;
 
@@ -58,10 +58,10 @@ public class ProductDAO extends AbstractDAO {
 
 			ResultSet rs = pstm.getGeneratedKeys();
 
-			int idProduct = 0;
+			long idProduct = 0;
 
 			if (rs.next()) {
-				idProduct = rs.getInt(1);
+				idProduct = rs.getLong(1);
 			}
 
 			return idProduct;
