+++ src/br/com/solstice/notecommerce/dao/impl/domain/user/customer/CreditCardDAO.java
@@ -23,7 +23,7 @@ public class CreditCardDAO extends AbstractDAO {
 	}
 
 	@Override
-	public int save(Entity entity) {
+	public long save(Entity entity) {
 		openConnection();
 		PreparedStatement pstm = null;
